Move RelocatedDwarf to symbol/relocate.rs
